app/requirements.txt
allure-pytest
boto3
moto
pika
pylint
pytest
